gst/vaapi: move JPEG encoder test classes to lib
